Created Account Library for frontend controller account.js for issue #228
